Add PEP registration
The PEP must be registered at the UCS to be able to get usage control requests evaluated.
The PEP sends a register request, and the UCS sends back a register response, indicating whether the registration was successful or not.
In case the PEP was already registered, the UCS replies that the registration succeeded, thus signalling to the PEP that it can send usage control requests.

If an unregistered PEP attempts a usage control request, that request is discarded and no answer is provided to the PEP.