Change usage-control image entrypoint to be used in ansible script
The run_UCSDht script mimics that of the one in the repository 'dht-to-mqtt', i.e., run_manager: basically, it checks if a particular file exists in the machine the container is running. If so, the main process (UCSDht.jar) is allowed to execute. If the file is not present in the machine, the process is not allowed to run.
This method relies on the fact that, at any time, the file is present on one machine only.