[UCS] Fix attributes retrieval from condition
The previous implementation did not allow for policies having an attribute with more than one attribute value.
For example, a policy with the function time-in-range uses two AttributeValue for the same AttributeDesignator.
In this case, the previous implementation creates two Attributes: the first is well populated with the attribute value and the info from the AttributeDesignator, but the second is populated only with the attribute value, leaving all the info derived from the AttributeDesignator not set.
On the contrary, the new implementation creates one Attribute, setting the info taken from the AttributeDesignator
and adding both the attribute values to the attributeValueMap of the Attribute.