Move the implementation of the majority of the methods from PIPReader to the abstract class PIPBase.
The PIPBase now uses a PIPSubscriberTimer, which is available to all the classes extending the PIPBase.
The PIPReaderSubscriberTimer has been removed.

In PIPReader (and future PIPs), the majority of the logic is the same.
The main method that has to be customized is the retrieve(attribute) method. Often, also the init(properties) method will be customized according to the PIP's task.

The UCSDhtPipReaderProperties class has been removed, and the UCSDhtPipProperties class has been edited in order to be used by all the future PIPs.
Custom additional properties of a PIP are stored in the map 'additionalProperties'.
For the PIPReader, the file name of an attribute is stored as value in that map for the key corresponding to the attributeId of the attribute it refers to.
Additionally, a default attribute value can be specified as an entry in the map, using the file name as key, and the attribute value as value.
For example, additionalProperties could be represented (in json notation) as:
additionalProperties: {
  "urn:...:attribute-id": "fileName.txt",
  "fileName.txt": "attributeValue"
}

Note that the attribute value is optional and not used by the PIPReader.
However, it can be useful if the PIP has to be serialized and later restored.

Edit PIPReader tests to comply with the new implementation.