Add PIPTime module
This PIP is intended to capture the current time.
It uses Balana as dependency to create a TimeAttribute object in order to encode the time in a format compliant with http://www.w3.org/2001/XMLSchema#time, used by Balana.

Note that this PIP does not use the additionalProperties map, and that the init method is not overridden.