Add AbstractPIPWebSocket
This abstract PIP is intended to be used as base class for those PIPs whose task is to retrieve data from the DHT.
It includes the strings dhtUri, topicName, and topicUuid.
These have to be defined as additional properties in the PipProperties object is passed to the constructor of the inheritor class.
It also includes a websocket client endpoint that connects to the dhtUri.
Finally, it implements the method performRequestGetTopicUuid() that queries the dht for the set topicName and topicUuid. This method returns a json string, which is the response obtained from the dht.