Use TypeAdapterFactory to register classes that implement RequestPostTopicUuid and Persistent
The DHTPersistentMessageClient needs to know how to deserialize a json string in order to understand whether the received string contains the right topicName and topicUuid.

Since now the RequestPostTopicUuid and Persistent classes are interfaces (in order to have more generality), the classes that implement them have to be registered to a TypeAdapterFactory, and the TypeAdapterFactory have to be registered to the GsonBuilder.

Moreover, when we intend to implement a new entity, e.g., a PIP, with a given 'topic name', we have to create the classes that implement RequestPostTopicUuid, Persistent (only if the entity is intended to upload data to the DHT), and another class implementing the fields contained in the 'value' field of the json, i.e., the fields we expect to find. This last class is implemented by Status and DomoLight in the current commit.